tests/wpt/web-platform-tests/html/canvas/offscreen/drawing-images-to-the-canvas/2d.drawImage.zerosource.image.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.drawImage.zerosource.image</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.drawImage.zerosource.image</h1>
<p class="desc">drawImage with zero-sized source rectangle from image draws nothing without exception</p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);
  var promise = new Promise(function(resolve, reject) {
    var xhr = new XMLHttpRequest();
    xhr.open("GET", '/images/red-zerowidth.svg');
    xhr.responseType = 'blob';
    xhr.send();
    xhr.onload = function() {
      resolve(xhr.response);
    };
  });
  promise.then(function(response) {
    createImageBitmap(response).then(bitmap => {
      ctx.drawImage(bitmap, 0, 0, 100, 50);
      _assertPixel(canvas, 50,25, 0,255,0,255);
    });
  });

}, "drawImage with zero-sized source rectangle from image draws nothing without exception");
</script>
